[UPDATE] The extensions file is refactored by dividing it by the classes to which the extensions belong.
