+++ app/src/main/java/com/softyorch/taskapp/ui/components/CheckCustom.kt
@@ -9,7 +9,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
-import com.softyorch.taskapp.utils.intOffsetAnimation
+import com.softyorch.taskapp.utils.extensions.intOffsetAnimation
 
 @ExperimentalMaterial3Api
 @Composable
