+++ app/src/main/java/com/softyorch/taskapp/ui/screens/detail/DetailScreen.kt
@@ -34,6 +34,7 @@ import com.softyorch.taskapp.ui.widgets.RowInfo
 import com.softyorch.taskapp.ui.widgets.ShowTask
 import com.softyorch.taskapp.ui.widgets.newTask.newTask
 import com.softyorch.taskapp.utils.*
+import com.softyorch.taskapp.utils.extensions.*
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import java.time.Instant
