+++ app/src/main/java/com/softyorch/taskapp/ui/screens/history/HistoryScreen.kt
@@ -29,6 +29,10 @@ import com.softyorch.taskapp.ui.components.dropDawnMenuCustom.dropDawnMenuCustom
 import com.softyorch.taskapp.ui.navigation.AppScreensRoutes
 import com.softyorch.taskapp.ui.widgets.RowInfo
 import com.softyorch.taskapp.utils.*
+import com.softyorch.taskapp.utils.extensions.containerColorAnimation
+import com.softyorch.taskapp.utils.extensions.intOffsetAnimationTransition
+import com.softyorch.taskapp.utils.extensions.toStringFormatDate
+import com.softyorch.taskapp.utils.extensions.toastError
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
