+++ app/src/main/java/com/softyorch/taskapp/utils/extensions/AnimationsBooleanExtensions.kt
@@ -1,9 +1,5 @@
-package com.softyorch.taskapp.utils
+package com.softyorch.taskapp.utils.extensions
 
-import android.annotation.SuppressLint
-import android.content.Context
-import android.widget.Toast
-import androidx.compose.animation.animateColor
 import androidx.compose.animation.animateColorAsState
 import androidx.compose.animation.core.*
 import androidx.compose.material3.MaterialTheme
@@ -11,43 +7,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.State
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.IntOffset
-import androidx.datastore.preferences.preferencesDataStore
-import java.security.SecureRandom
-import java.text.SimpleDateFormat
-import java.util.*
-
-@SuppressLint("SimpleDateFormat")
-fun Date.toStringFormatted(): String {
-    return this.toStringFormatDate() +
-            " at " +
-            this.toStringFormatHour()
-}
-
-@SuppressLint("SimpleDateFormat")
-fun Date.toStringFormatDate(): String =
-    SimpleDateFormat("yyyy-MM-dd").format(this)
-
-@SuppressLint("SimpleDateFormat")
-fun Date.toStringFormatHour(): String =
-    SimpleDateFormat("HH:mm:ss").format(this)
-
-@SuppressLint("SimpleDateFormat")
-fun String.toDate(): Date? =
-    SimpleDateFormat("EE MMM dd HH:mm:ss z yyyy", Locale.ENGLISH).parse(this)
-
-val Context.datastore by preferencesDataStore(name = USER_DATA)
-
-fun Context.toastError(message: String, onShow: () -> Unit) {
-    Toast.makeText(this, message, Toast.LENGTH_SHORT).show().also { onShow() }
-}
-
-fun Int.random(): Int {
-    val random = SecureRandom()
-    random.setSeed(random.generateSeed(this))
-    return random.nextInt(this - 1)
-}
-
-/** Animations ***************************************************/
+import com.softyorch.taskapp.utils.DURATION_MILLIS_BTN_CHANGE_COLOR
 
 /*@Composable
 fun Boolean.alphaAnimation(): State<Float> =
@@ -195,30 +155,3 @@ fun Boolean.intOffsetAnimationTransition(
         ),
         finishedListener = { finishedListener() }
     )
-
-@Composable
-fun MaterialTheme.infiniteTransitionAnimateColor(): State<Color> =
-    rememberInfiniteTransition().animateColor(
-        initialValue = this.colorScheme.tertiary,
-        targetValue = this.colorScheme.primary,
-        animationSpec = infiniteRepeatable(
-            animation = tween(200, easing = LinearEasing),
-            repeatMode = RepeatMode.Reverse
-        )
-    )
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
