+++ app/src/main/java/com/softyorch/taskapp/utils/extensions/DateExtensions.kt
@@ -1,0 +1,20 @@
+package com.softyorch.taskapp.utils.extensions
+
+import android.annotation.SuppressLint
+import java.text.SimpleDateFormat
+import java.util.*
+
+@SuppressLint("SimpleDateFormat")
+fun Date.toStringFormatted(): String {
+    return this.toStringFormatDate() +
+            " at " +
+            this.toStringFormatHour()
+}
+
+@SuppressLint("SimpleDateFormat")
+fun Date.toStringFormatDate(): String =
+    SimpleDateFormat("yyyy-MM-dd").format(this)
+
+@SuppressLint("SimpleDateFormat")
+fun Date.toStringFormatHour(): String =
+    SimpleDateFormat("HH:mm:ss").format(this)
