+++ app/src/main/java/com/softyorch/taskapp/utils/extensions/StringExtensions.kt
@@ -1,0 +1,9 @@
+package com.softyorch.taskapp.utils.extensions
+
+import android.annotation.SuppressLint
+import java.text.SimpleDateFormat
+import java.util.*
+
+@SuppressLint("SimpleDateFormat")
+fun String.toDate(): Date? =
+    SimpleDateFormat("EE MMM dd HH:mm:ss z yyyy", Locale.ENGLISH).parse(this)
